♻️ refactor: 서브 컴포넌트 코드 분리, 테스트 코드 작성
